backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.12
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
sqlalchemy==2.0.36
alembic==1.14.0
email-validator==2.3.0